Support Rails 7.1's Independent SchemaMigration object.
The SchemaMigration object no longer inherits from ActiveRecord::Base.

See "Move SchemaMigration to an independent object"
https://github.com/rails/rails/commit/436277da88507f9aae0874e62f3e61a8546b9683

Several methods were renamed, and are no longer available as class
methods on SchemaMigration, but as an instance of SchemaMigration
per connection.

The object per connection seems to remove the need to `reset_table_name`
in the tests. That method is not available on the new object.