Register connection adapter.
The initial connection creation method invocations were re-written to
reduce convention by adding an explicit registration mapping between
adapter name and class.

This implements two.
- adapter: oracle_enhanced
  This is the primary one that people will use.
- adapter: oracle
  This replaces the previous emulate_oracle_adapter database config
setting which we can no longer use, since the config is not available
when we are registering this adapter name to class mapping, but
accomplishes the same end. So if you have 'emulate_oracle_adapter: true`
in your configs, remove that, and set your `adapter: oracle`.

https://github.com/rails/rails/commit/009c7e74117690f0dbe200188a929b345c9306c1