write_lobs() is redundant
In the very distant past, LOBs were not possible to insert together with
the row. So this complicated callback method was necessary. Initially
an `empty_clob()` was inserted and then the callbacks would call
`write_lobs()` to insert the actually desired lobs.

But nowadays it is performing just a second redundant insert. With
prepared statements, the `OCI::CLOB/BLOB/NCLOB` objects are used. And
the LOBs get inserted from the beginning as they should. And this also
supports large LOBs.

In fact presently up to 2GB - 8 bytes but this can be improved even
further by making ruby-oci8 to use multiple `write` calls for the supplied
argument instead of a single one which fails for large values.

Not a high priority to increase max supported string size though, given
that is all this huge sting needs to be in memory, possibly multiple
copies.

This is not a loss of functionality compared to `write_lob()` because:
* `write_lobs()` also used a single `write` call
* if `OCI:BLOB` would crash then presence of `write_lobs()` wouldn't
  help